pyproject.toml
[tool.poetry]
name = "portfolio-tracker-google-sheets"
version = "0.1.0"
description = "A Python script that analyzes investment portfolio data from a Google Sheet"
authors = ["Adnan Hakim <adnanhakim007@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
gspread = "^6.1.0"
oauth2client = "^4.1.3"
requests = "^2.31.0"
xirr = "^0.1.8"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
